refactor mcp rules yaml helpers
